refactor: split model.go, add forward, remap keybindings
- Split model.go (5127→2105 lines) into 10 focused files:
  sidebar, messages, content, overlays, commands, save_attach,
  sync, update_ui, summary, util
- Add email forwarding (f key, NewForward, quoteForward)
  with original attachments and full forward headers
- Forward and Reply now work from message list, not just content pane
- Remap keybindings:
  s = sync (was f), ctrl+s = sync all
  f = forward (was w)
  v = AI summary (was s)
  ctrl+x = mark all read
- Rewrite install.sh: use GitHub redirect URL instead of API call,
  add gzip validation and tar error handling
- deploy.sh: make release notes optional in full_release()
