+++ internal/ui/help.go
@@ -48,9 +48,10 @@ func renderHelp(width int, styles Styles, keys KeyMap) string {
 				{keys.UnreadOnly.Help().Key, "toggle unread-only view"},
 				bind(keys.Archive),
 				{keys.Delete.Help().Key, "delete selected message"},
-			bind(keys.Compose),
-			bind(keys.Reply),
-			{keys.GrammarCheck.Help().Key, "AI grammar & spell check (in compose)"},
+				bind(keys.Compose),
+				bind(keys.Reply),
+				bind(keys.Forward),
+				{keys.GrammarCheck.Help().Key, "AI grammar & spell check (in compose)"},
 				{keys.ToggleHeaders.Help().Key, "toggle full email headers"},
 				{keys.SaveAttach.Help().Key, "save all attachments to folder"},
 				{keys.ToggleQuote.Help().Key, "toggle quoted text collapse"},
