+++ keybindings.csv
@@ -1,0 +1,54 @@
+GoField,Keys,HelpDisplay,Description,Context
+Up,"k,up",k/↑,up,global
+Down,"j,down",j/↓,down,global
+Left,"h,left",h/←,move left,global
+Right,"l,right",l/→,move right,global
+NextPane,"tab,]",tab/],next pane,global
+PrevPane,"shift+tab,[",shift+tab/[,prev pane,global
+Enter,enter,enter,open,global
+Back,esc,esc,back,global
+Sync,s,s,sync mailbox,main
+SyncAll,"F,ctrl+s",F/ctrl+s,sync all,main
+MarkRead,x,x,mark read,main
+MarkAllRead,"R,ctrl+x",R/ctrl+x,mark all read,main
+OpenBrowser,o,o,open in browser,main
+NextLink,"ctrl+n,alt+n",ctrl+n,next link,content
+PrevLink,"ctrl+p,alt+p",ctrl+p,prev link,content
+Search,/,/,search,main
+UnreadOnly,u,u,toggle unread only,main
+Compose,c,c,compose,main
+Reply,r,r,reply,main
+Forward,f,f,forward,main
+Archive,a,a,archive,main
+Command,"p,:",p,command palette,global
+AccountManager,m,m,accounts,global
+ThemePicker,T,T,theme picker,global
+Settings,S,S,settings,global
+UpdateInstall,U,U,install update,main
+UpdateIgnore,i,i,ignore update,main
+Help,?,?,help,global
+Quit,q,q,quit,global
+Summary,v,v,AI summary,main
+CopyText,C,C,copy text,summary
+SaveMD,M,M,save as .md,summary
+ContentSearch,ctrl+f,ctrl+f,find in content,content
+ToggleQuote,z,z,toggle quoted text,content
+ToggleHeaders,ctrl+e,ctrl+e,toggle full headers,content
+SaveAttach,ctrl+d,ctrl+d,save attachments,content
+Add,a,a,add account,account_manager
+Edit,e,e,edit account,account_manager
+Delete,d,d,delete,account_manager+main
+SaveAccount,"ctrl+s,f2",ctrl+s,save account,account_manager
+TestAccount,"ctrl+t,f5",ctrl+t,test account,account_manager
+AttachFile,ctrl+a,ctrl+a,attach file,compose
+GrammarCheck,ctrl+g,ctrl+g,AI grammar check,compose
+RemoveAttach,ctrl+r,ctrl+r,remove attachment,compose
+CycleSender,ctrl+u,ctrl+u,change sender,compose
+Confirm,enter,enter,confirm,overlay
+Cancel,esc,esc,cancel,overlay
+Space," ",space,toggle select,main
+SelectAll,A,A,select all,main
+Yes,y,y,yes,overlay
+No,n,n,no,overlay
+Tab,tab,tab,next field,overlay
+Backspace,backspace,backspace,delete,overlay
